+++ pkg/controller/app/checkout_state.go
@@ -10,24 +10,45 @@ import (
 	corev1obj "github.com/puppetlabs/leg/k8sutil/pkg/controller/obj/api/corev1"
 	"github.com/puppetlabs/leg/k8sutil/pkg/controller/obj/helper"
 	"github.com/puppetlabs/leg/k8sutil/pkg/controller/obj/lifecycle"
+	"github.com/puppetlabs/leg/k8sutil/pkg/norm"
 	pvpoolv1alpha1 "github.com/puppetlabs/pvpool/pkg/apis/pvpool.puppet.com/v1alpha1"
 	pvpoolv1alpha1obj "github.com/puppetlabs/pvpool/pkg/apis/pvpool.puppet.com/v1alpha1/obj"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
+const (
+	CheckoutReclaimPolicyAnnotationKey = "pvpool.puppet.com/checkout.reclaim-policy"
+)
+
 type CheckoutState struct {
-	Checkout              *pvpoolv1alpha1obj.Checkout
+	Checkout *pvpoolv1alpha1obj.Checkout
+
+	// PersistentVolumeClaim is the final, settled PVC that should be made
+	// available to the requestor.
 	PersistentVolumeClaim *corev1obj.PersistentVolumeClaim
 
 	// PersistentVolume is the PV corresponding to the PVC owned by this
 	// checkout. If this object exists, but not PersistentVolumeClaim, we merely
 	// need to set up the PVC to point at this PV.
 	PersistentVolume *corev1obj.PersistentVolume
+
+	// LockedPersistentVolumeClaim is a PVC that we use to take a PV from the
+	// pool for this checkout. This PVC is necessary if we need to perform
+	// modifications to the underlying PV before we can present it to the
+	// requestor (for example, setting fields on the volume source).
+	//
+	// Using this intermediate means the somewhat complicated logic to detect
+	// whether a pool replica has become stale can be left alone while we mess
+	// with the PV here instead.
+	LockedPersistentVolumeClaim *corev1obj.PersistentVolumeClaim
+
+	// LockedPersistentVolume is the PV corresponding to the locked PVC, i.e.,
+	// the original PV from the pool.
+	LockedPersistentVolume *corev1obj.PersistentVolume
 
 	// Conds represent status updates for given conditions.
 	Conds map[pvpoolv1alpha1.CheckoutConditionType]pvpoolv1alpha1.Condition
@@ -36,73 +57,7 @@ type CheckoutState struct {
 var _ lifecycle.Loader = &CheckoutState{}
 var _ lifecycle.Persister = &CheckoutState{}
 
-func (cs *CheckoutState) loadFromPVC(ctx context.Context, cl client.Client) error {
-	klog.V(4).InfoS("checkout state: load: loading from owned persistent volume claim", "checkout", cs.Checkout.Key, "pvc", cs.PersistentVolumeClaim.Key, "pv", cs.PersistentVolumeClaim.Object.Spec.VolumeName)
-
-	volume := corev1obj.NewPersistentVolume(cs.PersistentVolumeClaim.Object.Spec.VolumeName)
-	if ok, err := volume.Load(ctx, cl); err != nil {
-		return err
-	} else if !ok || volume.Object.Spec.ClaimRef == nil { // Not bound?
-		klog.InfoS("checkout state: load: persistent volume missing or not bound (not using)", "checkout", cs.Checkout.Key, "pvc", cs.PersistentVolumeClaim.Key, "pv", volume.Name)
-		return nil
-	}
-
-	cs.PersistentVolume = volume
-	return nil
-}
-
-func (cs *CheckoutState) loadFromVolumeName(ctx context.Context, cl client.Client) error {
-	volumeName := cs.Checkout.Object.Status.VolumeName
-	if volumeName == "" {
-		return nil
-	}
-
-	klog.V(4).InfoS("checkout state: load: loading from persistent volume name", "checkout", cs.Checkout.Key, "pv", volumeName)
-
-	volume := corev1obj.NewPersistentVolume(volumeName)
-	if ok, err := volume.Load(ctx, cl); err != nil {
-		return err
-	} else if !ok || volume.Object.Spec.ClaimRef == nil { // Not bound?
-		klog.InfoS("checkout state: load: persistent volume missing or not bound (not using)", "checkout", cs.Checkout.Key, "pv", volume.Name)
-		return nil
-	}
-
-	// Let's see which PVC this PV points at.
-	claim := corev1obj.NewPersistentVolumeClaim(client.ObjectKey{
-		Namespace: volume.Object.Spec.ClaimRef.Namespace,
-		Name:      volume.Object.Spec.ClaimRef.Name,
-	})
-	if ok, err := claim.Load(ctx, cl); err != nil {
-		return err
-	} else if !ok || !claim.Object.GetDeletionTimestamp().IsZero() {
-		// The claim was deleted. We need to pick a new one from the pool as we
-		// risk the volume moving to a Terminating state here.
-		klog.InfoS("checkout state: load: persistent volume claim is being deleted (not using)", "checkout", cs.Checkout.Key, "pvc", claim.Key, "pv", volume.Name)
-		return nil
-	}
-
-	if ctrl := metav1.GetControllerOf(claim.Object); ctrl != nil {
-		switch {
-		case ctrl.UID == cs.Checkout.Object.GetUID():
-			klog.V(4).InfoS("checkout state: load: persistent volume is used by this checkout", "checkout", cs.Checkout.Key, "pvc", claim.Key, "pv", volume.Name)
-		case schema.FromAPIVersionAndKind(ctrl.APIVersion, ctrl.Kind) == pvpoolv1alpha1obj.CheckoutKind:
-			// Race condition where the claim has been reassigned from under us.
-			klog.InfoS("checkout state: load: persistent volume stolen (not using)", "checkout", cs.Checkout.Key, "pvc", claim.Key, "pv", volume.Name)
-			return nil
-		default:
-			klog.V(4).InfoS("checkout state: load: persistent volume is still in pool", "checkout", cs.Checkout.Key, "pvc", claim.Key, "pv", volume.Name)
-		}
-	}
-
-	cs.PersistentVolume = volume
-	return nil
-}
-
 func (cs *CheckoutState) loadFromPool(ctx context.Context, cl client.Client) (bool, error) {
-	if cs.PersistentVolume != nil {
-		return true, nil
-	}
-
 	namespace := cs.Checkout.Object.Spec.PoolRef.Namespace
 	if namespace == "" {
 		namespace = cs.Checkout.Key.Namespace
@@ -147,84 +102,141 @@ func (cs *CheckoutState) loadFromPool(ctx context.Context, cl client.Client) (bo
 		pr := ps.Available[0]
 
 		klog.V(4).InfoS("checkout state: load: using PVC from pool", "checkout", cs.Checkout.Key, "pool", pool.Key, "pvc", pr.PersistentVolumeClaim.Key, "pv", pr.PersistentVolume.Name)
-		cs.PersistentVolume = pr.PersistentVolume
+		cs.LockedPersistentVolume = pr.PersistentVolume
 	}
 
 	return true, nil
 }
 
 func (cs *CheckoutState) Load(ctx context.Context, cl client.Client) (bool, error) {
-	if _, err := cs.PersistentVolumeClaim.Load(ctx, cl); err != nil {
-		return false, nil
+	pairs := []struct {
+		PersistentVolumeClaim *corev1obj.PersistentVolumeClaim
+		PersistentVolume      **corev1obj.PersistentVolume
+	}{
+		{
+			PersistentVolumeClaim: cs.LockedPersistentVolumeClaim,
+			PersistentVolume:      &cs.LockedPersistentVolume,
+		},
+		{
+			PersistentVolumeClaim: cs.PersistentVolumeClaim,
+			PersistentVolume:      &cs.PersistentVolume,
+		},
 	}
+	for _, pair := range pairs {
+		if ok, err := pair.PersistentVolumeClaim.Load(ctx, cl); err != nil {
+			return false, err
+		} else if !ok {
+			continue
+		}
 
-	if helper.Exists(cs.PersistentVolumeClaim.Object) {
-		if ctrl := metav1.GetControllerOf(cs.PersistentVolumeClaim.Object); ctrl == nil || ctrl.UID != cs.Checkout.Object.GetUID() {
+		if ctrl := metav1.GetControllerOf(pair.PersistentVolumeClaim.Object); ctrl == nil || ctrl.UID != cs.Checkout.Object.GetUID() {
 			cs.Conds[pvpoolv1alpha1.CheckoutAcquired] = pvpoolv1alpha1.Condition{
 				Status:  corev1.ConditionUnknown,
 				Reason:  pvpoolv1alpha1.CheckoutAcquiredReasonConflict,
-				Message: fmt.Sprintf("A non-controlled PVC with the name %s already exists.", cs.PersistentVolumeClaim.Key.Name),
+				Message: fmt.Sprintf("A non-controlled PVC with the name %s already exists.", pair.PersistentVolumeClaim.Key.Name),
 			}
 			return false, errmark.MarkTransient(fmt.Errorf("a PVC with this checkout's desired name already exists"))
 		}
+
+		volume := corev1obj.NewPersistentVolume(pair.PersistentVolumeClaim.Object.Spec.VolumeName)
+		if ok, err := volume.Load(ctx, cl); err != nil {
+			return false, err
+		} else if !ok {
+			continue
+		}
+
+		// Sanity check: volume must cross-reference the PVC for successful
+		// pre-bind.
+		if volume.Object.Spec.ClaimRef == nil || volume.Object.Spec.ClaimRef.UID != pair.PersistentVolumeClaim.Object.GetUID() {
+			klog.InfoS("checkout state: load: persistent volume does not match claim (not using)", "checkout", cs.Checkout.Key, "pvc", pair.PersistentVolumeClaim.Key, "pv", volume.Name)
+			continue
+		}
+
+		*pair.PersistentVolume = volume
 	}
 
-	switch cs.PersistentVolumeClaim.Object.Status.Phase {
-	case corev1.ClaimPending, corev1.ClaimBound:
-		if err := cs.loadFromPVC(ctx, cl); err != nil {
-			return false, err
+	if cs.PersistentVolumeClaim.Object.Status.Phase != corev1.ClaimBound && cs.LockedPersistentVolume == nil {
+		// We need to get one from the pool.
+		if ok, err := cs.loadFromPool(ctx, cl); err != nil || !ok {
+			return ok, err
 		}
-	default:
-		// Either this object doesn't yet exist or the claim was lost for some
-		// reason.
-		if err := cs.loadFromVolumeName(ctx, cl); err != nil {
-			return false, err
-		}
-	}
-
-	// If we did not end up setting the PV above, we need to request a new one
-	// from the pool.
-	if ok, err := cs.loadFromPool(ctx, cl); err != nil || !ok {
-		return ok, err
 	}
 
 	return true, nil
 }
 
 func (cs *CheckoutState) Persist(ctx context.Context, cl client.Client) error {
-	if err := cs.Checkout.Own(ctx, cs.PersistentVolumeClaim); err != nil {
-		return err
-	}
-
-	if err := cs.PersistentVolumeClaim.Persist(ctx, cl); errors.IsInvalid(err) {
-		cs.Conds[pvpoolv1alpha1.CheckoutAcquired] = pvpoolv1alpha1.Condition{
-			Status:  corev1.ConditionFalse,
-			Reason:  pvpoolv1alpha1.CheckoutAcquiredReasonInvalid,
-			Message: fmt.Sprintf("The PVC could not be created because of configuration problems: %v", err),
-		}
-		return errmark.MarkUser(err)
-	} else if err != nil {
-		return err
-	}
-
-	if cs.PersistentVolume == nil {
-		eventctx.EventRecorder(ctx).Event(cs.Checkout.Object, "Warning", "VolumeAttachment", "No volume found")
-
-		return errmark.MarkTransient(fmt.Errorf("missing persistent volume"))
-	}
-
-	// Sync UID to ClaimRef.
-	cs.PersistentVolume.Object.Spec.ClaimRef.UID = cs.PersistentVolumeClaim.Object.GetUID()
-
-	if err := cs.PersistentVolume.Persist(ctx, cl); err != nil {
-		return err
-	}
-
-	if cs.PersistentVolumeClaim.Object.Status.Phase == corev1.ClaimBound {
+	// We can either be in a state where we're still trying to allocate the PV
+	// and PVC, or we can have the PVC settled and bound. If it's not bound,
+	// we'll also set up the locked PV/PVC.
+	switch cs.PersistentVolumeClaim.Object.Status.Phase {
+	case corev1.ClaimBound:
 		cs.Conds[pvpoolv1alpha1.CheckoutAcquired] = pvpoolv1alpha1.Condition{
 			Status:  corev1.ConditionTrue,
 			Reason:  pvpoolv1alpha1.CheckoutAcquiredReasonCheckedOut,
 			Message: "The PVC is ready to use.",
+		}
+
+		// Now that we've allocated everything, we no longer need the locked PV
+		// (must be manually deleted) or PVC.
+		if cs.LockedPersistentVolume != nil {
+			if _, err := cs.LockedPersistentVolume.Delete(ctx, cl, lifecycle.DeleteWithPropagationPolicy(metav1.DeletePropagationBackground)); err != nil {
+				return err
+			}
+		}
+
+		if _, err := cs.LockedPersistentVolumeClaim.Delete(ctx, cl); err != nil {
+			return err
+		}
+	default:
+		// Note on the logic here: if we successfully wrote both the locked PVC
+		// and PV, we have completed the pre-bind and we're guaranteed to have
+		// it settle at some point.
+		//
+		// We don't actually need to wait for the controller to make that
+		// happen, but if it does, we'll just go through the reconcile loop
+		// again.
+		pairs := []struct {
+			PersistentVolumeClaim *corev1obj.PersistentVolumeClaim
+			PersistentVolume      *corev1obj.PersistentVolume
+		}{
+			{
+				PersistentVolumeClaim: cs.LockedPersistentVolumeClaim,
+				PersistentVolume:      cs.LockedPersistentVolume,
+			},
+			{
+				PersistentVolumeClaim: cs.PersistentVolumeClaim,
+				PersistentVolume:      cs.PersistentVolume,
+			},
+		}
+		for _, pair := range pairs {
+			if err := cs.Checkout.Own(ctx, pair.PersistentVolumeClaim); err != nil {
+				return err
+			}
+
+			if err := pair.PersistentVolumeClaim.Persist(ctx, cl); errors.IsInvalid(err) {
+				cs.Conds[pvpoolv1alpha1.CheckoutAcquired] = pvpoolv1alpha1.Condition{
+					Status:  corev1.ConditionFalse,
+					Reason:  pvpoolv1alpha1.CheckoutAcquiredReasonInvalid,
+					Message: fmt.Sprintf("The PVC %q could not be created because of configuration problems: %v", pair.PersistentVolumeClaim.Key.Name, err),
+				}
+				return errmark.MarkUser(err)
+			} else if err != nil {
+				return err
+			}
+
+			if pair.PersistentVolume == nil {
+				eventctx.EventRecorder(ctx).Event(cs.Checkout.Object, "Warning", "VolumeAttachment", fmt.Sprintf("Volume for PVC %q could not be found", pair.PersistentVolumeClaim.Key.Name))
+
+				return errmark.MarkTransient(fmt.Errorf("missing persistent volume"))
+			}
+
+			// Sync UID to ClaimRef.
+			pair.PersistentVolume.Object.Spec.ClaimRef.UID = pair.PersistentVolumeClaim.Object.GetUID()
+
+			if err := pair.PersistentVolume.Persist(ctx, cl); err != nil {
+				return err
+			}
 		}
 	}
 
@@ -243,17 +255,41 @@ func NewCheckoutState(c *pvpoolv1alpha1obj.Checkout) *CheckoutState {
 			Namespace: c.Key.Namespace,
 			Name:      claimName,
 		}),
+		LockedPersistentVolumeClaim: corev1obj.NewPersistentVolumeClaim(client.ObjectKey{
+			Namespace: c.Key.Namespace,
+			Name:      norm.MetaNameSuffixed(c.Key.Name, "-locked"),
+		}),
 		Conds: make(map[pvpoolv1alpha1.CheckoutConditionType]pvpoolv1alpha1.Condition),
 	}
 }
 
 func ConfigureCheckoutState(cs *CheckoutState) (*CheckoutState, error) {
-	if cs.PersistentVolume == nil {
+	switch {
+	case cs.PersistentVolumeClaim.Object.Status.Phase == corev1.ClaimBound:
+		return cs, nil
+	case cs.LockedPersistentVolume == nil:
 		klog.V(4).InfoS("checkout state: configure: no volume, clearing status", "checkout", cs.Checkout.Key)
 		return cs, nil
+	case cs.PersistentVolume == nil:
+		// Uniqueness follows the underlying volume.
+		cs.PersistentVolume = corev1obj.NewPersistentVolume(norm.MetaNameSuffixed("pvpool", "-"+string(cs.LockedPersistentVolume.Object.GetUID())))
 	}
 
-	// Set up the PV to point at our claim.
+	// We need to keep track of the original reclaim policy so we can use it for
+	// our new PV. We use an annotation to do so.
+	//
+	// It's important that we only set this once, because once we modify the
+	// target PV, it will not have the right value!
+	if _, ok := cs.LockedPersistentVolume.Object.GetAnnotations()[CheckoutReclaimPolicyAnnotationKey]; !ok {
+		helper.Annotate(cs.LockedPersistentVolume.Object, CheckoutReclaimPolicyAnnotationKey, string(cs.LockedPersistentVolume.Object.Spec.PersistentVolumeReclaimPolicy))
+	}
+
+	// Copy locked PV to new PV. Note that we also copy annotations as they are
+	// used to keep track of deallocators in CSI.
+	helper.CopyLabelsAndAnnotations(cs.PersistentVolume.Object, cs.LockedPersistentVolume.Object)
+	cs.LockedPersistentVolume.Object.Spec.DeepCopyInto(&cs.PersistentVolume.Object.Spec)
+
+	// Set up PV.
 	cs.PersistentVolume.Object.Spec.AccessModes = cs.Checkout.Object.Spec.AccessModes
 	cs.PersistentVolume.Object.Spec.ClaimRef = &corev1.ObjectReference{
 		APIVersion: corev1obj.PersistentVolumeClaimKind.GroupVersion().String(),
@@ -261,6 +297,15 @@ func ConfigureCheckoutState(cs *CheckoutState) (*CheckoutState, error) {
 		Namespace:  cs.PersistentVolumeClaim.Key.Namespace,
 		Name:       cs.PersistentVolumeClaim.Key.Name,
 		UID:        cs.PersistentVolumeClaim.Object.GetUID(),
+	}
+	cs.PersistentVolume.Object.Spec.PersistentVolumeReclaimPolicy = corev1.PersistentVolumeReclaimPolicy(cs.LockedPersistentVolume.Object.GetAnnotations()[CheckoutReclaimPolicyAnnotationKey])
+
+	// In most volume sources, setting the readOnly field just forces mounts to
+	// also have the readOnly flag set. However, for CSI, it also changes the
+	// behavior of the external attacher so that it requests the correct mode
+	// from the CSI driver.
+	if spec := &cs.PersistentVolume.Object.Spec; spec.CSI != nil {
+		spec.CSI.ReadOnly = len(spec.AccessModes) == 1 && spec.AccessModes[0] == corev1.ReadOnlyMany
 	}
 
 	// Set up the PVC to point at the PV.
@@ -270,6 +315,29 @@ func ConfigureCheckoutState(cs *CheckoutState) (*CheckoutState, error) {
 	cs.PersistentVolumeClaim.Object.Spec.Resources = corev1.ResourceRequirements{
 		Requests: corev1.ResourceList{
 			corev1.ResourceStorage: cs.PersistentVolume.Object.Spec.Capacity.Storage().DeepCopy(),
+		},
+	}
+
+	// Set up locked PV to point at our locked PVC.
+	cs.LockedPersistentVolume.Object.Spec.ClaimRef = &corev1.ObjectReference{
+		APIVersion: corev1obj.PersistentVolumeClaimKind.GroupVersion().String(),
+		Kind:       corev1obj.PersistentVolumeClaimKind.Kind,
+		Namespace:  cs.LockedPersistentVolumeClaim.Key.Namespace,
+		Name:       cs.LockedPersistentVolumeClaim.Key.Name,
+		UID:        cs.LockedPersistentVolumeClaim.Object.GetUID(),
+	}
+
+	// Set locked PV to retain so we won't release the underlying storage when
+	// we delete the locked PVC.
+	cs.LockedPersistentVolume.Object.Spec.PersistentVolumeReclaimPolicy = corev1.PersistentVolumeReclaimRetain
+
+	// Configure the temporary locked PVC.
+	cs.LockedPersistentVolumeClaim.Object.Spec.StorageClassName = &cs.LockedPersistentVolume.Object.Spec.StorageClassName
+	cs.LockedPersistentVolumeClaim.Object.Spec.VolumeName = cs.LockedPersistentVolume.Name
+	cs.LockedPersistentVolumeClaim.Object.Spec.AccessModes = cs.LockedPersistentVolume.Object.Spec.AccessModes
+	cs.LockedPersistentVolumeClaim.Object.Spec.Resources = corev1.ResourceRequirements{
+		Requests: corev1.ResourceList{
+			corev1.ResourceStorage: cs.LockedPersistentVolume.Object.Spec.Capacity.Storage().DeepCopy(),
 		},
 	}